+++ RELEASE_NOTES.md
@@ -10,3 +10,5 @@ v0.68.0
 - Node: The InputStream abstraction now has an optional close() method, which will be called by MessageBackup.validate() on any created streams when the operation is complete. If your InputStream already has a close() method, this may be a breaking change for you.
 
 - backups: Enforce that messages with expiration timers < 24 hours are not included in Remote Backups.
+
+- backups: Add support for LocalLocator for local backups